TASK: Move form helpers from Domain/Model to eel
