+++ Classes/Eel/AbstractFormHelper.php
@@ -1,7 +1,17 @@
 <?php
 declare(strict_types=1);
 
-namespace Neos\Fusion\Form\Domain\Model;
+namespace Neos\Fusion\Form\Eel;
+
+/*
+ * This file is part of the Neos.Fusion.Form package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
 
 use Neos\Eel\ProtectedContextAwareInterface;
